# docs/memx_spec_v3/docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: memx v1 API Contract
  version: 1.0.0
  description: |
    Canonical contract for v1 required endpoints.
    Requirement mapping is embedded via x-requirement-id.
servers:
  - url: http://127.0.0.1:7766
paths:
  /v1/notes:ingest:
    post:
      operationId: notesIngest
      x-requirement-id: REQ-API-001
      summary: Ingest a new note.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotesIngestRequest'
      responses:
        '200':
          description: Note was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotesIngestResponse'
        '400':
          $ref: '#/components/responses/InvalidArgumentError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/notes:search:
    post:
      operationId: notesSearch
      x-requirement-id: REQ-API-001
      summary: Search notes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotesSearchRequest'
      responses:
        '200':
          description: Search results.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotesSearchResponse'
        '400':
          $ref: '#/components/responses/InvalidArgumentError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/notes/{id}:
    get:
      operationId: notesGet
      x-requirement-id: REQ-API-001
      summary: Get one note by id.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
          x-requirement-id: REQ-API-001
      responses:
        '200':
          description: Requested note.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '400':
          $ref: '#/components/responses/InvalidArgumentError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/gc:run:
    post:
      operationId: gcRun
      x-requirement-id: REQ-GC-001
      summary: Run short-store GC (experimental).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GCRunRequest'
      responses:
        '200':
          description: GC completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GCRunResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          description: Feature disabled fallback (`mem.features.gc_short=false`) or internal error.
          x-requirement-id: REQ-ERR-001
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
                  - type: object
                    properties:
                      code:
                        const: INTERNAL
                    description: |
                      When /v1/gc:run is published but mem.features.gc_short is OFF,
                      return INTERNAL as the canonical v1 fallback (FAILED_PRECONDITION-equivalent).
  /v1/docs:ingest:
    post:
      operationId: docsIngest
      summary: Ingest a resolver document and generate chunks.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocsIngestRequest'
      responses:
        '200':
          description: Document was ingested.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocsIngestResponse'
        '400':
          $ref: '#/components/responses/InvalidArgumentError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/docs:resolve:
    post:
      operationId: docsResolve
      summary: Resolve required and recommended documents.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocsResolveRequest'
      responses:
        '200':
          description: Resolved document sets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocsResolveResponse'
        '400':
          $ref: '#/components/responses/InvalidArgumentError'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/chunks:get:
    post:
      operationId: chunksGet
      summary: Get document chunks and prompt-ready memory cards.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChunksGetRequest'
      responses:
        '200':
          description: Requested chunks.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChunksGetResponse'
        '400':
          $ref: '#/components/responses/InvalidArgumentError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/docs:search:
    post:
      operationId: docsSearch
      summary: Search resolver documents with optional structured filters.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocsSearchRequest'
      responses:
        '200':
          description: Search results.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocsSearchResponse'
        '400':
          $ref: '#/components/responses/InvalidArgumentError'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/cards:search:
    post:
      operationId: cardsSearch
      summary: Search prompt-ready memory cards.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardsSearchRequest'
      responses:
        '200':
          description: Ranked memory cards.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardsSearchResponse'
        '400':
          $ref: '#/components/responses/InvalidArgumentError'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/cards:feedback:
    post:
      operationId: cardsFeedback
      summary: Record actual memory card usage for adaptive ranking.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardFeedbackRequest'
      responses:
        '200':
          description: Feedback was recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardFeedbackResponse'
        '400':
          $ref: '#/components/responses/InvalidArgumentError'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/cards:bundle:
    post:
      operationId: cardsBundle
      summary: Export memory cards as a prompt-ready bundle.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromptBundleRequest'
      responses:
        '200':
          description: Prompt-ready memory card bundle.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromptBundleResponse'
        '400':
          $ref: '#/components/responses/InvalidArgumentError'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/taskstate:export:
    post:
      operationId: taskStateExport
      summary: Export resolver read receipts, required docs, and stale reasons for agent-taskstate.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskStateExportRequest'
      responses:
        '200':
          description: agent-taskstate-friendly resolver state export.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStateExportResponse'
        '400':
          $ref: '#/components/responses/InvalidArgumentError'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/reads:ack:
    post:
      operationId: readsAck
      summary: Acknowledge that a task read a document version.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReadsAckRequest'
      responses:
        '200':
          description: Read receipt was recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadsAckResponse'
        '400':
          $ref: '#/components/responses/InvalidArgumentError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/docs:stale-check:
    post:
      operationId: docsStaleCheck
      summary: Check whether a task has stale read receipts.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocsStaleCheckRequest'
      responses:
        '200':
          description: Stale status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocsStaleCheckResponse'
        '400':
          $ref: '#/components/responses/InvalidArgumentError'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/contracts:resolve:
    post:
      operationId: contractsResolve
      summary: Resolve contract material for a feature or task.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractsResolveRequest'
      responses:
        '200':
          description: Resolved contract material.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractsResolveResponse'
        '400':
          $ref: '#/components/responses/InvalidArgumentError'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  responses:
    InvalidArgumentError:
      description: Invalid request.
      x-requirement-id: REQ-ERR-001
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  code:
                    const: INVALID_ARGUMENT
    NotFoundError:
      description: Resource not found.
      x-requirement-id: REQ-ERR-001
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  code:
                    const: NOT_FOUND
    ConflictError:
      description: Request conflict.
      x-requirement-id: REQ-ERR-001
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  code:
                    const: CONFLICT
...
    InternalError:
      description: Internal error.
      x-requirement-id: REQ-ERR-001
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  code:
                    const: INTERNAL
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        details:
          description: Optional machine-readable details.
      x-requirement-id: REQ-ERR-001
    Note:
      type: object
      additionalProperties: false
      required:
        - id
        - title
        - summary
        - body
        - created_at
        - updated_at
        - last_accessed_at
        - access_count
        - source_type
        - origin
        - source_trust
        - sensitivity
      properties:
        id:
          type: string
        title:
          type: string
        summary:
          type: string
        body:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
        last_accessed_at:
          type: string
        access_count:
          type: integer
          format: int64
        source_type:
          type: string
        origin:
          type: string
        source_trust:
          type: string
        sensitivity:
          type: string
      x-requirement-id: REQ-API-001
    NotesIngestRequest:
      type: object
      additionalProperties: false
      required: [title, body]
      properties:
        title:
          type: string
        body:
          type: string
        summary:
          type: string
        source_type:
          type: string
        origin:
          type: string
        source_trust:
          type: string
        sensitivity:
          type: string
        tags:
          type: array
          items:
            type: string
      x-requirement-id: REQ-API-001
    NotesIngestResponse:
      type: object
      additionalProperties: false
      required: [note]
      properties:
        note:
          $ref: '#/components/schemas/Note'
      x-requirement-id: REQ-API-001
    NotesSearchRequest:
      type: object
      additionalProperties: false
      required: [query]
      properties:
        query:
          type: string
        top_k:
          type: integer
      x-requirement-id: REQ-API-001
    NotesSearchResponse:
      type: object
      additionalProperties: false
      required: [notes]
      properties:
        notes:
          type: array
          items:
            $ref: '#/components/schemas/Note'
      x-requirement-id: REQ-API-001
...
    DocsIngestRequest:
      type: object
      additionalProperties: false
      required: [doc_type, title, version, body]
      properties:
        doc_id:
          type: string
        doc_type:
          type: string
        title:
          type: string
        source_path:
          type: string
        version:
          type: string
        updated_at:
          type: string
        tags:
          type: array
          items:
            type: string
        feature_keys:
          type: array
          items:
            type: string
        task_ids:
          type: array
          items:
            type: string
        summary:
          type: string
        body:
          type: string
        chunking:
          $ref: '#/components/schemas/ChunkingOptions'
        acceptance_criteria:
          type: array
          items:
            type: string
        forbidden_patterns:
          type: array
          items:
            type: string
        definition_of_done:
          type: array
          items:
            type: string
        dependencies:
          type: array
          items:
            type: string
    DocsIngestResponse:
      type: object
      additionalProperties: false
      required: [doc_id, version, chunk_count, status]
      properties:
        doc_id:
          type: string
        version:
          type: string
        chunk_count:
          type: integer
        status:
          type: string
          enum: [ingested]
    DocsResolveRequest:
      type: object
      additionalProperties: false
      properties:
        feature:
          type: string
        task_id:
          type: string
        topic:
          type: string
        limit:
          type: integer
    DocsResolveResponse:
      type: object
      additionalProperties: false
      required: [required, recommended]
      properties:
        required:
          type: array
          items:
            $ref: '#/components/schemas/ResolveEntry'
        recommended:
          type: array
          items:
            $ref: '#/components/schemas/ResolveEntry'
    ChunksGetRequest:
      type: object
      additionalProperties: false
      properties:
        doc_id:
          type: string
        query:
          type: string
        heading:
          type: string
        limit:
          type: integer
        chunk_ids:
          type: array
          items:
            type: string
    ChunksGetResponse:
      type: object
      additionalProperties: false
      required: [doc_id, chunks]
      properties:
        doc_id:
          type: string
        chunks:
          type: array
          items:
            $ref: '#/components/schemas/ResolverChunk'
        memory_cards:
          type: array
          items:
            $ref: '#/components/schemas/ResolverMemoryCard'
    DocsSearchRequest:
      type: object
      additionalProperties: false
      required: [query]
      properties:
        query:
          type: string
        doc_types:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        feature_keys:
          type: array
          items:
            type: string
        limit:
          type: integer
    DocsSearchResponse:
      type: object
      additionalProperties: false
      required: [results]
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/ResolveEntry'
    CardsSearchRequest:
      type: object
      additionalProperties: false
      required: [query]
      properties:
        query:
          type: string
        doc_types:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        feature_keys:
          type: array
          items:
            type: string
        memory_types:
          type: array
          items:
            $ref: '#/components/schemas/ResolverMemoryType'
        limit:
          type: integer
        token_budget:
          type: integer
        ranking_weights:
          $ref: '#/components/schemas/MemoryCardRankingWeights'
    CardsSearchResponse:
      type: object
      additionalProperties: false
      required: [cards]
      properties:
        cards:
          type: array
          items:
            $ref: '#/components/schemas/ResolverMemoryCard'
...
    CardFeedbackRequest:
      type: object
      additionalProperties: false
      required: [card_id, signal]
      properties:
        card_id:
          type: string
        doc_id:
          type: string
        chunk_id:
          type: string
        memory_type:
          type: string
        signal:
          type: string
          enum: [used, helpful, pinned, irrelevant, skipped]
        weight:
          type: integer
        query:
          type: string
...
    CardFeedbackResponse:
      type: object
      additionalProperties: false
      required: [feedback]
      properties:
        feedback:
          $ref: '#/components/schemas/CardFeedbackRecord'
...
    TaskStateExportRequest:
      type: object
      additionalProperties: false
      required: [task_id]
      properties:
        task_id:
          type: string
        feature:
          type: string
...
    ReadsAckResponse:
      type: object
      additionalProperties: false
      required: [receipt]
      properties:
        receipt:
          $ref: '#/components/schemas/ResolverReadReceipt'
...
    GCRunRequest:
      type: object
      additionalProperties: false
      required: [target]
      properties:
        target:
          type: string
        options:
          type: object
          additionalProperties: false
          properties:
            dry_run:
              type: boolean
      x-requirement-id: REQ-GC-001
    GCRunResponse:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
      x-requirement-id: REQ-GC-001